+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**Settings**) Add KOReader sync settings
 - (**Settings**) Add database settings
 - (**Settings**) Add hikari connection pool database setting
+- (**Settings**) Add OPDS CBZ MIME-Type setting
 - (**Browse**) Add "open in webview" button in source browse page
 
 ### Changed